@@ -1,14 +1,12 @@
-﻿using Core.Entities;
-using Core.Entities.Patients;
-using System;
+﻿using Core.Entities.Patients;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Core.Interfaces
 {
     public interface IPatientRepository : IAsyncRepository<Patient>
     {
+        Task<Patient> GetPatientsTest();
         Task<List<PatientBirthStatistik>> GetGeburtsDatumCountByYears();
     }
 }
