@@ -1,18 +1,18 @@
-﻿using Core.Interfaces;
-using Infrastructure.Repositories;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.AspNetCore.Identity;
-using Core.Entities.Identity;
-using Infrastructure.Data;
+﻿//using Core.Interfaces;
+//using Infrastructure.Repositories;
+//using Microsoft.EntityFrameworkCore;
+//using Microsoft.Extensions.DependencyInjection;
+//using Microsoft.AspNetCore.Identity;
+//using Core.Entities.Identity;
+//using Infrastructure.Data;
 
-namespace Infrastructure.Configurations
-{
-    public static class ServiceCollectionExtensions
-    {
-        public static void AddDataAccessServices(this IServiceCollection services, string connectionString)
-        {
-            
-        }
-    }
-}
+//namespace Infrastructure.Configurations
+//{
+//    public static class ServiceCollectionExtensions
+//    {
+//        public static void AddDataAccessServices(this IServiceCollection services, string connectionString)
+//        {
+
+//        }
+//    }
+//}
