@@ -1,26 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
-using Core.Entities.Identity;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Threading.Tasks;
+//using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+//using Microsoft.EntityFrameworkCore;
+//using Core.Entities.Identity;
 
-namespace Infrastructure.Data
-{
-    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
-    {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
-            : base(options)
-        {
-        }
+//namespace Infrastructure.Data
+//{
+//    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+//    {
+//        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options)
+//            : base(options)
+//        {
+//        }
 
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
-            // Customize the ASP.NET Identity model and override the defaults if needed.
-            // For example, you can rename the ASP.NET Identity table names and more.
-            // Add your customizations after calling base.OnModelCreating(builder);
-        }
-    }
-}
+//        protected override void OnModelCreating(ModelBuilder builder)
+//        {
+//            base.OnModelCreating(builder);
+//            // Customize the ASP.NET Identity model and override the defaults if needed.
+//            // For example, you can rename the ASP.NET Identity table names and more.
+//            // Add your customizations after calling base.OnModelCreating(builder);
+//        }
+//    }
+//}
