@@ -1,217 +1,217 @@
-﻿// <auto-generated />
-using Cabinet.Data;
-using Infrastructure.Repositories;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Metadata;
-using Microsoft.EntityFrameworkCore.Migrations;
-using System;
+﻿//// <auto-generated />
+//using Cabinet.Data;
+//using Infrastructure.Repositories;
+//using Microsoft.EntityFrameworkCore;
+//using Microsoft.EntityFrameworkCore.Infrastructure;
+//using Microsoft.EntityFrameworkCore.Metadata;
+//using Microsoft.EntityFrameworkCore.Migrations;
+//using System;
 
-namespace Infrastructure.Data.Migrations
-{
-    [DbContext(typeof(ApplicationDbContext))]
-    partial class ApplicationDbContextModelSnapshot : ModelSnapshot
-    {
-        protected override void BuildModel(ModelBuilder modelBuilder)
-        {
-#pragma warning disable 612, 618
-            modelBuilder
-                .HasAnnotation("ProductVersion", "1.0.0-rc3")
-                .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
+//namespace Infrastructure.Data.Migrations
+//{
+//    [DbContext(typeof(ApplicationDbContext))]
+//    partial class ApplicationDbContextModelSnapshot : ModelSnapshot
+//    {
+//        protected override void BuildModel(ModelBuilder modelBuilder)
+//        {
+//#pragma warning disable 612, 618
+//            modelBuilder
+//                .HasAnnotation("ProductVersion", "1.0.0-rc3")
+//                .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
-            modelBuilder.Entity("Cabinet.Models.ApplicationUser", b =>
-                {
-                    b.Property<string>("Id");
+//            modelBuilder.Entity("Cabinet.Models.ApplicationUser", b =>
+//                {
+//                    b.Property<string>("Id");
 
-                    b.Property<int>("AccessFailedCount");
+//                    b.Property<int>("AccessFailedCount");
 
-                    b.Property<string>("ConcurrencyStamp")
-                        .IsConcurrencyToken();
+//                    b.Property<string>("ConcurrencyStamp")
+//                        .IsConcurrencyToken();
 
-                    b.Property<string>("Email")
-                        .HasMaxLength(256);
+//                    b.Property<string>("Email")
+//                        .HasMaxLength(256);
 
-                    b.Property<bool>("EmailConfirmed");
+//                    b.Property<bool>("EmailConfirmed");
 
-                    b.Property<bool>("LockoutEnabled");
+//                    b.Property<bool>("LockoutEnabled");
 
-                    b.Property<DateTimeOffset?>("LockoutEnd");
+//                    b.Property<DateTimeOffset?>("LockoutEnd");
 
-                    b.Property<string>("NormalizedEmail")
-                        .HasMaxLength(256);
+//                    b.Property<string>("NormalizedEmail")
+//                        .HasMaxLength(256);
 
-                    b.Property<string>("NormalizedUserName")
-                        .HasMaxLength(256);
+//                    b.Property<string>("NormalizedUserName")
+//                        .HasMaxLength(256);
 
-                    b.Property<string>("PasswordHash");
+//                    b.Property<string>("PasswordHash");
 
-                    b.Property<string>("PhoneNumber");
+//                    b.Property<string>("PhoneNumber");
 
-                    b.Property<bool>("PhoneNumberConfirmed");
+//                    b.Property<bool>("PhoneNumberConfirmed");
 
-                    b.Property<string>("SecurityStamp");
+//                    b.Property<string>("SecurityStamp");
 
-                    b.Property<bool>("TwoFactorEnabled");
+//                    b.Property<bool>("TwoFactorEnabled");
 
-                    b.Property<string>("UserName")
-                        .HasMaxLength(256);
+//                    b.Property<string>("UserName")
+//                        .HasMaxLength(256);
 
-                    b.HasKey("Id");
+//                    b.HasKey("Id");
 
-                    b.HasIndex("NormalizedEmail")
-                        .HasName("EmailIndex");
+//                    b.HasIndex("NormalizedEmail")
+//                        .HasName("EmailIndex");
 
-                    b.HasIndex("NormalizedUserName")
-                        .IsUnique()
-                        .HasName("UserNameIndex");
+//                    b.HasIndex("NormalizedUserName")
+//                        .IsUnique()
+//                        .HasName("UserNameIndex");
 
-                    b.ToTable("AspNetUsers");
-                });
+//                    b.ToTable("AspNetUsers");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRole", b =>
-                {
-                    b.Property<string>("Id");
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRole", b =>
+//                {
+//                    b.Property<string>("Id");
 
-                    b.Property<string>("ConcurrencyStamp")
-                        .IsConcurrencyToken();
+//                    b.Property<string>("ConcurrencyStamp")
+//                        .IsConcurrencyToken();
 
-                    b.Property<string>("Name")
-                        .HasMaxLength(256);
+//                    b.Property<string>("Name")
+//                        .HasMaxLength(256);
 
-                    b.Property<string>("NormalizedName")
-                        .HasMaxLength(256);
+//                    b.Property<string>("NormalizedName")
+//                        .HasMaxLength(256);
 
-                    b.HasKey("Id");
+//                    b.HasKey("Id");
 
-                    b.HasIndex("NormalizedName")
-                        .HasName("RoleNameIndex");
+//                    b.HasIndex("NormalizedName")
+//                        .HasName("RoleNameIndex");
 
-                    b.ToTable("AspNetRoles");
-                });
+//                    b.ToTable("AspNetRoles");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRoleClaim<string>", b =>
-                {
-                    b.Property<int>("Id")
-                        .ValueGeneratedOnAdd();
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRoleClaim<string>", b =>
+//                {
+//                    b.Property<int>("Id")
+//                        .ValueGeneratedOnAdd();
 
-                    b.Property<string>("ClaimType");
+//                    b.Property<string>("ClaimType");
 
-                    b.Property<string>("ClaimValue");
+//                    b.Property<string>("ClaimValue");
 
-                    b.Property<string>("RoleId")
-                        .IsRequired();
+//                    b.Property<string>("RoleId")
+//                        .IsRequired();
 
-                    b.HasKey("Id");
+//                    b.HasKey("Id");
 
-                    b.HasIndex("RoleId");
+//                    b.HasIndex("RoleId");
 
-                    b.ToTable("AspNetRoleClaims");
-                });
+//                    b.ToTable("AspNetRoleClaims");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserClaim<string>", b =>
-                {
-                    b.Property<int>("Id")
-                        .ValueGeneratedOnAdd();
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserClaim<string>", b =>
+//                {
+//                    b.Property<int>("Id")
+//                        .ValueGeneratedOnAdd();
 
-                    b.Property<string>("ClaimType");
+//                    b.Property<string>("ClaimType");
 
-                    b.Property<string>("ClaimValue");
+//                    b.Property<string>("ClaimValue");
 
-                    b.Property<string>("UserId")
-                        .IsRequired();
+//                    b.Property<string>("UserId")
+//                        .IsRequired();
 
-                    b.HasKey("Id");
+//                    b.HasKey("Id");
 
-                    b.HasIndex("UserId");
+//                    b.HasIndex("UserId");
 
-                    b.ToTable("AspNetUserClaims");
-                });
+//                    b.ToTable("AspNetUserClaims");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserLogin<string>", b =>
-                {
-                    b.Property<string>("LoginProvider");
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserLogin<string>", b =>
+//                {
+//                    b.Property<string>("LoginProvider");
 
-                    b.Property<string>("ProviderKey");
+//                    b.Property<string>("ProviderKey");
 
-                    b.Property<string>("ProviderDisplayName");
+//                    b.Property<string>("ProviderDisplayName");
 
-                    b.Property<string>("UserId")
-                        .IsRequired();
+//                    b.Property<string>("UserId")
+//                        .IsRequired();
 
-                    b.HasKey("LoginProvider", "ProviderKey");
+//                    b.HasKey("LoginProvider", "ProviderKey");
 
-                    b.HasIndex("UserId");
+//                    b.HasIndex("UserId");
 
-                    b.ToTable("AspNetUserLogins");
-                });
+//                    b.ToTable("AspNetUserLogins");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserRole<string>", b =>
-                {
-                    b.Property<string>("UserId");
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserRole<string>", b =>
+//                {
+//                    b.Property<string>("UserId");
 
-                    b.Property<string>("RoleId");
+//                    b.Property<string>("RoleId");
 
-                    b.HasKey("UserId", "RoleId");
+//                    b.HasKey("UserId", "RoleId");
 
-                    b.HasIndex("RoleId");
+//                    b.HasIndex("RoleId");
 
-                    b.HasIndex("UserId");
+//                    b.HasIndex("UserId");
 
-                    b.ToTable("AspNetUserRoles");
-                });
+//                    b.ToTable("AspNetUserRoles");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserToken<string>", b =>
-                {
-                    b.Property<string>("UserId");
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserToken<string>", b =>
+//                {
+//                    b.Property<string>("UserId");
 
-                    b.Property<string>("LoginProvider");
+//                    b.Property<string>("LoginProvider");
 
-                    b.Property<string>("Name");
+//                    b.Property<string>("Name");
 
-                    b.Property<string>("Value");
+//                    b.Property<string>("Value");
 
-                    b.HasKey("UserId", "LoginProvider", "Name");
+//                    b.HasKey("UserId", "LoginProvider", "Name");
 
-                    b.ToTable("AspNetUserTokens");
-                });
+//                    b.ToTable("AspNetUserTokens");
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRoleClaim<string>", b =>
-                {
-                    b.HasOne("Microsoft.AspNetCore.Identity.IdentityRole")
-                        .WithMany("Claims")
-                        .HasForeignKey("RoleId")
-                        .OnDelete(DeleteBehavior.Cascade);
-                });
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityRoleClaim<string>", b =>
+//                {
+//                    b.HasOne("Microsoft.AspNetCore.Identity.IdentityRole")
+//                        .WithMany("Claims")
+//                        .HasForeignKey("RoleId")
+//                        .OnDelete(DeleteBehavior.Cascade);
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserClaim<string>", b =>
-                {
-                    b.HasOne("Cabinet.Models.ApplicationUser")
-                        .WithMany("Claims")
-                        .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
-                });
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserClaim<string>", b =>
+//                {
+//                    b.HasOne("Cabinet.Models.ApplicationUser")
+//                        .WithMany("Claims")
+//                        .HasForeignKey("UserId")
+//                        .OnDelete(DeleteBehavior.Cascade);
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserLogin<string>", b =>
-                {
-                    b.HasOne("Cabinet.Models.ApplicationUser")
-                        .WithMany("Logins")
-                        .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
-                });
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserLogin<string>", b =>
+//                {
+//                    b.HasOne("Cabinet.Models.ApplicationUser")
+//                        .WithMany("Logins")
+//                        .HasForeignKey("UserId")
+//                        .OnDelete(DeleteBehavior.Cascade);
+//                });
 
-            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserRole<string>", b =>
-                {
-                    b.HasOne("Microsoft.AspNetCore.Identity.IdentityRole")
-                        .WithMany("Users")
-                        .HasForeignKey("RoleId")
-                        .OnDelete(DeleteBehavior.Cascade);
+//            modelBuilder.Entity("Microsoft.AspNetCore.Identity.IdentityUserRole<string>", b =>
+//                {
+//                    b.HasOne("Microsoft.AspNetCore.Identity.IdentityRole")
+//                        .WithMany("Users")
+//                        .HasForeignKey("RoleId")
+//                        .OnDelete(DeleteBehavior.Cascade);
 
-                    b.HasOne("Cabinet.Models.ApplicationUser")
-                        .WithMany("Roles")
-                        .HasForeignKey("UserId")
-                        .OnDelete(DeleteBehavior.Cascade);
-                });
-#pragma warning restore 612, 618
-        }
-    }
-}
+//                    b.HasOne("Cabinet.Models.ApplicationUser")
+//                        .WithMany("Roles")
+//                        .HasForeignKey("UserId")
+//                        .OnDelete(DeleteBehavior.Cascade);
+//                });
+//#pragma warning restore 612, 618
+//        }
+//    }
+//}
